Узоры, убирание костылей.

+++ include/effects.h
@@ -1443,10 +1443,10 @@ private:
     byte csum = 0;
     byte _bri = 255U;
     byte buff[20 * 20];
-    uint32_t xsin, ysin;
+    float xsin, ysin;
     unsigned long lastrun2;
     byte _sc = 0;
-    int8_t _speedX, _speedY;
+    float _speedX, _speedY;
 
     CHSV colorMR[12] = {
         CHSV(0, 0, 0),              // 0 - Black
